backend/requirements.txt
fastapi
pydantic[email]
python-dotenv
psycopg2-binary
sqlalchemy
passlib[bcrypt]
bcrypt==3.2.0
python-jose[cryptography]
email-validator
python-multipart
httpx
google-generativeai
